requirements_minimal.txt
# Core Flask dependencies
Flask==3.1.2
flask-cors==6.0.1
pymongo==4.15.3
python-dotenv==1.1.1
requests==2.32.5

# Basic AI/NLP (no compilation required)
textblob==0.17.1
nltk==3.8.1

# Basic utilities
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2

# Certificate generation
reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0

# Security
cryptography==41.0.8
python-jose==3.3.0
bcrypt==4.1.2
